+++ examples/06_wim_info/main.go
@@ -12,7 +12,7 @@ import (
 	"os"
 	"text/tabwriter"
 
-	"github.com/deploymenttheory/go-sdk-windowsuup/windowsuup/wim"
+	"github.com/deploymenttheory/go-sdk-windowsuup/pkg/wim"
 )
 
 func main() {